Revert "Eval still does not work... Trying to change the model to use 1 input now"
This reverts commit d74fa784334c46da1fd05b0d9ac7f861424bcc26.

+++ src/engine/Evaluate.h
@@ -9,7 +9,7 @@ public:
 
     // returns true if position a is better for white than position b,
     // and false if the opposite
-    bool evaluate(std::vector<int> a, std::vector<int> b);
+    bool evaluate(std::vector<uint8_t> a, std::vector<uint8_t> b);
 
 private:
     torch::jit::script::Module module;
